% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown_assemble.R
\name{markdown_assemble}
\alias{markdown_assemble}
\title{Assemble Markdown Files into a Larger Document}
\usage{
markdown_assemble(title, author, directory)
}
\arguments{
\item{title}{The title of the Larger Document.}

\item{author}{The author of the Larger Document.}

\item{directory}{The path to the directory containing the Markdown Files.}
}
\value{
A Larger Document- e.g. a HTML web page- with the content of the Markdown Files.
}
\description{
Assemble Markdown Files into a Larger Document
}
