% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdownAssembleR-package.R
\docType{package}
\name{markdownAssembleR-package}
\alias{markdownAssembleR}
\alias{markdownAssembleR-package}
\title{markdownAssembleR: Assemble Markdown Files into Larger Documents}
\description{
Concatenates multiple markdown files together and renders a larger document (e.g. a PDF using LaTeX, or a HTML page).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rnnh/markdownAssembleR}
  \item Report bugs at \url{https://github.com/rnnh/markdownAssembleR/issues}
}

}
\author{
\strong{Maintainer}: Ronan Harrington \email{ronan_harr@protonmail.com} (\href{https://orcid.org/0000-0003-0616-907X}{ORCID})

}
\keyword{internal}
